requirements.txt
pyzmq-static==2.2
pyzmq==15.2.0
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.3
